change LONG to INTEGER for table column types

+++ lib/lib-table-cluster/src/test/java/org/sagebionetworks/table/cluster/TableIndexDAOImplTest.java
@@ -387,7 +387,7 @@ public class TableIndexDAOImplTest {
 		foo.setId("111");
 		foo.setMaximumSize(10L);
 		ColumnModel bar = new ColumnModel();
-		bar.setColumnType(ColumnType.LONG);
+		bar.setColumnType(ColumnType.INTEGER);
 		bar.setId("222");
 		bar.setName("bar");
 		List<ColumnModel> schema = new LinkedList<ColumnModel>();
@@ -436,7 +436,7 @@ public class TableIndexDAOImplTest {
 		foo.setId("111");
 		foo.setMaximumSize(10L);
 		ColumnModel bar = new ColumnModel();
-		bar.setColumnType(ColumnType.LONG);
+		bar.setColumnType(ColumnType.INTEGER);
 		bar.setId("222");
 		bar.setName("bar");
 		List<ColumnModel> schema = new LinkedList<ColumnModel>();
